Refactor parser for app settings
